Change Map Part in MainActivity to Fragment
